Disable resolve step on merge if .gitattributes isn't present.
Modified tools/merge_upstream.sh to not attempt to resolve conflicts if
a .gitattributes file isn't present.  If no .gitattributes file is present
the script exits gracefully rather than throwing an error when trying to read
the file.

Tested:
$ git checkout -b test goog/ub-games-master
$ git rm -rf Box2D/Build
$ git commit -m 'Removed Box2D/Build.'
$ ./tools/merge_upstream.sh
... lots of output ...
Automatic merge failed; fix conflicts and then commit the result.
U       Box2D/Build/vs2012/Box2D.sln
U       Box2D/Build/vs2012/Box2D.vcxproj
U       Box2D/Build/vs2012/Box2D.vcxproj.filters
U       Box2D/Build/vs2012/FreeGLUT.vcxproj
U       Box2D/Build/vs2012/FreeGLUT.vcxproj.filters
U       Box2D/Build/vs2012/GLUI.vcxproj
U       Box2D/Build/vs2012/GLUI.vcxproj.filters
U       Box2D/Build/vs2012/HelloWorld.vcxproj
U       Box2D/Build/vs2012/HelloWorld.vcxproj.filters
U       Box2D/Build/vs2012/Testbed.vcxproj
U       Box2D/Build/vs2012/Testbed.vcxproj.filters
U       Box2D/Build/xcode5/Box2D.xcodeproj/project.pbxproj
U       Box2D/Build/xcode5/Box2D.xcodeproj/project.xcworkspace/contents.xcworkspacedata
error: 'commit' is not possible because you have unmerged files.
hint: Fix them up in the work tree,
hint: and then use 'git add/rm <file>' as
hint: appropriate to mark resolution and make a commit,
hint: or use 'git commit -a'.
fatal: Exiting because of an unresolved conflict.

+++ liquidfun/tools/merge_upstream.sh
@@ -96,35 +96,37 @@ add_fetch_git_remote() {
 # have been deleted in our branch and added / copied / deleted / renamed in
 # another so they're handled here.
 resolve_deleted_renamed_conflicts() {
-  (
-    # Disable filename expansion (globbing) so that it's possible to compare
-    # each filename to a glob in the if statement below.
-    set -f
-    IFS=$'\n'
-	local merge_ours=$(awk '
-      /^#/ {
-        next;
-      }
+  if [[ -e .gitattributes ]]; then
+    (
+      # Disable filename expansion (globbing) so that it's possible to compare
+      # each filename to a glob in the if statement below.
+      set -f
+      IFS=$'\n'
+      local merge_ours=$(awk '
+        /^#/ {
+          next;
+        }
 
-      $2 ~ /merge=ours/ {
-        filename_glob = $1;
-        # Escape for bash if statement below.
-        gsub(/\//, "\\/", filename_glob);
-        gsub(/\*/, "*", filename_glob);
-        print filename_glob;
-      }' .gitattributes)
-    for file in $(git status -s | \
-                  awk '$1 ~ /^U[ACDR]/ || $1 ~ /^DU/ { print $2 }'); do
-      for use_ours_glob in ${merge_ours}; do
-        eval "
-          if [[ \"\${file/${use_ours_glob}/}\" == \"\" ]]; then
-            echo 'Dropping upstream changes to ${file}' >&2
-            git rm ${file}
-          fi
-        "
+        $2 ~ /merge=ours/ {
+          filename_glob = $1;
+          # Escape for bash if statement below.
+          gsub(/\//, "\\/", filename_glob);
+          gsub(/\*/, "*", filename_glob);
+          print filename_glob;
+        }' .gitattributes)
+      for file in $(git status -s | \
+                    awk '$1 ~ /^U[ACDR]/ || $1 ~ /^DU/ { print $2 }'); do
+        for use_ours_glob in ${merge_ours}; do
+          eval "
+            if [[ \"\${file/${use_ours_glob}/}\" == \"\" ]]; then
+              echo 'Dropping upstream changes to ${file}' >&2
+              git rm ${file}
+            fi
+          "
+        done
       done
-    done
-  )
+    )
+  fi
 }
 
 main() {
@@ -152,7 +154,7 @@ main() {
   # Mirror the subversion repo into a local git repository.
   if [[ "${SVN_UPSTREAM}" != "" ]]; then
     mirror_svn "${temp_upstream}" "${SVN_UPSTREAM}"
-	GIT_UPSTREAM="${temp_upstream}"
+    GIT_UPSTREAM="${temp_upstream}"
   fi
 
   if [[ "${GIT_UPSTREAM}" == "" ]] ;then
